Decrease buf_t size

+++ src/buffer.h
@@ -6,7 +6,7 @@
 
 typedef struct buf_s {
     uint8_t *bytes;
-    size_t size;
+    uint16_t size;
 } buf_t;
 
-bool parse_to_sized_buffer(uint8_t *in_buffer, size_t in_size, buf_t *out, size_t *offset);
+bool parse_to_sized_buffer(uint8_t *in_buffer, uint16_t in_size, buf_t *out, uint16_t *offset);
